# internal/docs/spec/swagger.yaml
basePath: /api/v1
definitions:
...
    github_com_Angle-HR_server_internal_apidoc.Meta:
        properties:
            has_more:
                type: boolean
            next_cursor:
                type: string
            request_id:
                example: abc123
                type: string
        type: object
...
    internal_handler.AdminUserDetailEnvelope:
        properties:
            data:
                $ref: '#/definitions/internal_handler.adminUserDetail'
            meta:
                $ref: '#/definitions/github_com_Angle-HR_server_internal_apidoc.Meta'
        type: object
    internal_handler.AdminUserListEnvelope:
        properties:
            data:
                items:
                    $ref: '#/definitions/internal_handler.adminUserListItem'
                type: array
            meta:
                $ref: '#/definitions/github_com_Angle-HR_server_internal_apidoc.Meta'
        type: object
    internal_handler.AdminUserPatchRequest:
        properties:
            deleted:
                type: boolean
        type: object
...
    internal_handler.adminUserDetail:
        properties:
            account_type:
                type: string
            completed_steps:
                items:
                    type: string
                type: array
            country_id:
                type: string
            created_at:
                type: string
            deleted_at:
                type: string
            email:
                type: string
            email_verified_at:
                type: string
            first_name:
                type: string
            last_name:
                type: string
            legal_full_name:
                type: string
            onboarding_completed_at:
                type: string
            onboarding_step:
                type: string
            organization_legal_name:
                type: string
            region:
                type: string
            region_source:
                type: string
            registry_id:
                type: string
            updated_at:
                type: string
            user_id:
                type: string
        type: object
    internal_handler.adminUserListItem:
        properties:
            created_at:
                type: string
            email:
                type: string
            id:
                type: string
            region:
                type: string
            updated_at:
                type: string
            user_id:
                type: string
        type: object
...
host: localhost:8080
info:
    contact: {}
    description: |-
        Waitlist and product onboarding API for Open HR.
        Interactive docs (Scalar) are served at `/` in non-production environments.
        Product flow: signup → verify email (6-digit OTP, 5 min expiry) → profile → address → [business] → complete.
        Onboarding step endpoints require `Authorization: Bearer <access_token>` after email verification.
        Admin ops console APIs live under `/api/v1/admin` and require a separate admin JWT (`token_use=admin`).
    title: Angle HR API
    version: "1.0"
paths:
...
    /admin/users:
        get:
            parameters:
                - description: Region filter
                  in: query
                  name: region
                  type: string
                - description: Search email
                  in: query
                  name: q
                  type: string
                - description: Page size
                  in: query
                  name: limit
                  type: integer
                - description: Offset
                  in: query
                  name: offset
                  type: integer
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/internal_handler.AdminUserListEnvelope'
            security:
                - BearerAuth: []
            summary: List product users
            tags:
                - admin/users
    /admin/users/{id}:
        get:
            parameters:
                - description: Registry UUID or regional user UUID
                  in: path
                  name: id
                  required: true
                  type: string
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/internal_handler.AdminUserDetailEnvelope'
            security:
                - BearerAuth: []
            summary: Get product user detail
            tags:
                - admin/users
        patch:
            consumes:
                - application/json
            description: 'Soft-delete or restore a regional account (`deleted: true|false`).'
            parameters:
                - description: Registry UUID
                  in: path
                  name: id
                  required: true
                  type: string
                - description: Patch payload
                  in: body
                  name: body
                  required: true
                  schema:
                    $ref: '#/definitions/internal_handler.AdminUserPatchRequest'
            produces:
                - application/json
            responses:
                "200":
                    description: OK
                    schema:
                        $ref: '#/definitions/internal_handler.AdminUserDetailEnvelope'
            security:
                - BearerAuth: []
            summary: Update product user
            tags:
                - admin/users
...
securityDefinitions:
    BearerAuth:
        description: JWT access token. Send the raw token, or optionally prefix with "Bearer ".
        in: header
        name: Authorization
        type: apiKey
swagger: "2.0"
...